feat: séparer entrainement & diagnostic
