+++ IA/README.md
@@ -1,3 +1,3 @@
-# Projet-Big-Data-2026
+# Projet-BigData-IA-Web-2026
 
 Projet sur le patrimoine arboré de St-Quentin
